Main header navigation steps
Initial writing of the main_header_navigation_steps.  This is a rough
method that uses a manual entry of each link name.  I am fairly certain
that there should be a more elegant way that will allow me to loop, but
I am not sure if that is consistent with Cucumber methodology.